GSdx: the promised index buffer update, needed a lot of changes, expect bugs in the next dozen revisions.
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@5045 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/GSdx/GSVector.h
@@ -1900,7 +1900,7 @@ public:
 		d = f.uph64(d);
 	}
 
-	__forceinline static bool compare16(const void* dst, const void* src, int size)
+	__forceinline static bool compare16(const void* dst, const void* src, size_t size)
 	{
 		ASSERT((size & 15) == 0);
 
@@ -1909,7 +1909,7 @@ public:
 		GSVector4i* s = (GSVector4i*)src;
 		GSVector4i* d = (GSVector4i*)dst;
 
-		for(int i = 0; i < size; i++)
+		for(size_t i = 0; i < size; i++)
 		{
 			if(!d[i].eq(s[i]))
 			{
@@ -1920,7 +1920,7 @@ public:
 		return true;
 	}
 
-	__forceinline static bool compare64(const void* dst, const void* src, int size)
+	__forceinline static bool compare64(const void* dst, const void* src, size_t size)
 	{
 		ASSERT((size & 63) == 0);
 
@@ -1929,7 +1929,7 @@ public:
 		GSVector4i* s = (GSVector4i*)src;
 		GSVector4i* d = (GSVector4i*)dst;
 
-		for(int i = 0; i < size; i += 4)
+		for(size_t i = 0; i < size; i += 4)
 		{
 			GSVector4i v0 = (d[i * 4 + 0] == s[i * 4 + 0]);
 			GSVector4i v1 = (d[i * 4 + 1] == s[i * 4 + 1]);
@@ -1948,7 +1948,7 @@ public:
 		return true;
 	}
 
-	__forceinline static bool update(const void* dst, const void* src, int size)
+	__forceinline static bool update(const void* dst, const void* src, size_t size)
 	{
 		ASSERT((size & 15) == 0);
 
@@ -1959,7 +1959,7 @@ public:
 
 		GSVector4i v = GSVector4i::xffffffff();
 
-		for(int i = 0; i < size; i++)
+		for(size_t i = 0; i < size; i++)
 		{
 			v &= d[i] == s[i];
